State Scripts

+++ scripts/states.py
@@ -1,30 +1,53 @@
 # Created by Yam for the Russia Reworked mod. Free to use.
-
 import os
 import json
 import sys
 import re
 import yaml
+import traceback
 from datetime import date
 
 if __name__ == '__main__':
 
     try:
-        # Put this file in the root folder of HOI4 along with run_states.bat and simply run it
-        filepath = sys.path[0]
+        # Either put this in the root folder of HOI4 or put the full path to your HOI4 folder
         directory = r".\history\states"
         localisation = r".\localisation\state_names_l_english.yml"
-        fullcwd = os.chdir(filepath + directory)
+        filepath = input("Enter the path to your HOI4 installation folder (%s) : " % sys.path[0])
+        if not os.path.isdir(filepath):
+            print("Path given (%s) is not a valid directory" % filepath)
+            print("Cancelling...")
+            sys.exit()
+
+        if filepath == "":
+            filepath = sys.path[0]
+
+        fullcwd = filepath + directory
         localisationpath = filepath + localisation
 
+        if not os.path.isdir(fullcwd):
+            print("Cannot find folder \"states\" (%s) " % fullcwd)
+            print("Cancelling...")
+            sys.exit()
+
+        if not os.path.isfile(localisationpath):
+            print("Cannot find file \"state_names_l_english.yml\" (%s) " % localisationpath)
+            print("Cancelling...")
+            sys.exit()
+
+        fullcwd = os.chdir(fullcwd)
         statecount = 0
         data = {
             "statecount": statecount
         }
 
         with open(localisationpath, 'r', encoding='utf-8-sig') as yml:
-            statelocalisation = yaml.load(yml)
-            statelocalisation = statelocalisation['l_english'].split()
+            yml = yml.read()
+            yml = yml.replace(":0", ":")
+            yml = yml.replace(":1", ":")
+            yml = yml.replace(":2", ":")
+            statelocalisation = yaml.safe_load(yml)
+            statelocalisation = statelocalisation['l_english']
 
         for filename in os.listdir(fullcwd):
             statecount += 1
@@ -75,13 +98,10 @@ if __name__ == '__main__':
 
                         provincefound = False
 
-                if name + ":0" in statelocalisation:
-                    nameloca = statelocalisation[statelocalisation.index(name + ":0") + 1]
-                elif name + ":1" in statelocalisation:
-                    nameloca = statelocalisation[statelocalisation.index(name + ":1") + 1]
-                elif name + ":2" in statelocalisation:
-                    nameloca = statelocalisation[statelocalisation.index(name + ":2") + 1]
-                nameloca = nameloca.strip("\"")
+                for key, value in statelocalisation.items():
+                    if key == name:
+                        nameloca = value
+                        break
 
             state = {
                 "filename": filename,
@@ -104,15 +124,25 @@ if __name__ == '__main__':
 
         print("%s states added" % statecount)
 
-        os.chdir(sys.path[0])
-
-        jsonfilename = "states_%s.json" % date.today()
-        with open(jsonfilename, 'w', encoding='utf-8-sig') as f:
-            json.dump(data, f, ensure_ascii=False, indent=4)
-
         print("Success")
-        print("Saved as %s" % jsonfilename)
+
+        os.chdir(sys.path[0])
+        jsonfilename = "states_%s.json" % date.today()
+
+        if os.path.isfile(jsonfilename):
+            overwrite = input("Warning : File %s already exist. Replace? (yes/no) : " % jsonfilename)
+            if overwrite == "yes" or overwrite == "y" or overwrite == "":
+                with open(jsonfilename, 'w', encoding='utf-8-sig') as f:
+                    json.dump(data, f, ensure_ascii=False, indent=4)
+
+                print("Saved as %s" % jsonfilename)
+        else:
+            with open(jsonfilename, 'w', encoding='utf-8-sig') as f:
+                json.dump(data, f, ensure_ascii=False, indent=4)
+
+            print("Saved as %s" % jsonfilename)
     
-    except:
-        print("An error occurred")
+    except Exception:
+        print("\nAn error occurred :", end="\n\n")
+        print(traceback.format_exc())
         print("Cancelling...")
